+++ docs/source/autoapi/markupsafe/_speedups/index.rst
@@ -1,0 +1,19 @@
+markupsafe._speedups
+====================
+
+.. py:module:: markupsafe._speedups
+
+
+
+
+Module Contents
+---------------
+
+.. py:function:: escape(s)
+
+.. py:function:: escape_silent(s)
+
+.. py:function:: soft_str(s)
+
+.. py:function:: soft_unicode(s)
+
