+++ docs/source/autoapi/rpaudio/rpaudio/index.rst
@@ -1,0 +1,6 @@
+rpaudio.rpaudio
+===============
+
+.. py:module:: rpaudio.rpaudio
+
+
