+++ docs/source/autoapi/websockets/speedups/index.rst
@@ -1,0 +1,13 @@
+websockets.speedups
+===================
+
+.. py:module:: websockets.speedups
+
+
+
+
+Module Contents
+---------------
+
+.. py:function:: apply_mask(data, mask)
+
