+++ examples/AudioChannel/single_channel_example.py
@@ -1,4 +1,4 @@
-import rpaudio
+import python.rpaudio.rpaudio as rpaudio
 import asyncio
 from datetime import datetime, timedelta
