+++ examples/AudioSink/single_audio_sink_example.py
@@ -1,4 +1,4 @@
-import rpaudio
+import python.rpaudio.rpaudio as rpaudio
 import asyncio
